Support simple #define constants
Also introduce 'macroDefined' lexer state that signals that there's at
least one macro definition and we must take slower code path and test
each standalone name, because it may be a macro that we should
substitute. This looks pretty complex, but it does not seem like
there's a simpler way that's just as performant.